feat(Xor): Link Xor with the parsing
I linked the Xor with the Parsing to start the xor in the terminal
