+++ Parser.py
@@ -119,6 +119,8 @@ class Parser:
         self.message = sys.stdin.read().strip()
         if not self.message:
             raise Error("The message is empty")
+        if self.system == "xor" and self.hasOption and len(self.key) == len(self.message):
+            return
         if self.hasOption and len(self.key) != len(self.message) * 2:
             raise Error("The key length must be equal to the message length because the \"-b\" flag is used")
